+++ lib/view_model/cubit/photo_details_cubit/photodetails_state.dart
@@ -6,3 +6,5 @@ abstract class PhotodetailsState {}
 class PhotodetailsInitial extends PhotodetailsState {}
 
 class PhotodetailsLoaded extends PhotodetailsState {}
+
+class Details extends PhotodetailsState {}
